LayoutTests/js/parse-backslash-before-newline-expected.txt

PASS "teststring with CR LF" is "teststring with CR LF"
PASS "test
string with LF CR" threw exception SyntaxError: Unexpected EOF.
PASS "teststring with CR" is "teststring with CR"
PASS "teststring with LF" is "teststring with LF"
PASS successfullyParsed is true

TEST COMPLETE
